# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendor/orders:
    get:
      operationId: listVendorOrders
      tags: [orders]
      summary: List order line items for the current user's own vendor(s)
      responses:
        "200":
          description: The vendor's order line items, newest order first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorOrderItemList"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
...
  /checkout:
    post:
      operationId: checkout
      tags: [orders]
      summary: Convert the current session's cart into a real order
      description: |
        For `momo` orders, first charges the buyer's mobile money wallet
        through Paystack for the quoted cart total; the order is only
        created if that charge succeeds. For `cash_on_delivery` orders, no
        charge is attempted and the order is confirmed immediately with a
        `cod_pending` payment event. Either way, runs the checkout workflow
        in a single DB transaction: validates the cart is non-empty and the
        address belongs to the caller, reserves inventory per line item
        (rejecting on insufficient available stock), optionally validates
        and applies a promotion code to the order total, creates the order +
        order items + a fulfillment record per vendor + a payment event,
        clears the cart, then emits an `order.placed` domain event on
        success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: The created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Cart is empty, the promotion code is invalid/expired/not met, or the address does not belong to the caller
        "401":
          description: Authentication required
        "402":
          description: The mobile money charge was declined, not approved, or could not be completed
        "409":
          description: Insufficient stock for one or more items
...
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List the current buyer's own orders, newest first
      responses:
        "200":
          description: The current user's orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
        "401":
          description: Authentication required
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order the caller can access (own order, own vendor's line items, or admin/support_agent)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [orders]
      summary: List all orders (admin/support_agent only)
      responses:
        "200":
          description: All orders, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
        "401":
          description: Authentication required
        "403":
          description: Forbidden
...
components:
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, fulfilled, cancelled]
    OrderPaymentMethod:
      type: string
      enum: [momo, cash_on_delivery]
    MomoProvider:
      type: string
      enum: [mtn, vodafone, airteltigo]
    OrderItem:
      type: object
      required: [id, orderId, productId, vendorId, titleSnapshot, pricePesewasSnapshot, qty, subtotalPesewas]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        productId: { type: integer }
        vendorId: { type: integer }
        titleSnapshot: { type: string }
        pricePesewasSnapshot: { type: integer }
        qty: { type: integer }
        subtotalPesewas: { type: integer }
    Order:
      type: object
      required: [id, buyerUserId, status, subtotalPesewas, discountPesewas, totalPesewas, createdAt, items, fulfillments]
      properties:
        id: { type: integer }
        buyerUserId: { type: integer }
        status:
          $ref: "#/components/schemas/OrderStatus"
        subtotalPesewas: { type: integer }
        discountPesewas: { type: integer }
        promotionCode: { type: ["string", "null"] }
        totalPesewas: { type: integer }
        addressId: { type: ["integer", "null"] }
        paymentMethod:
          oneOf:
            - $ref: "#/components/schemas/OrderPaymentMethod"
            - type: "null"
        address:
          description: The delivery address snapshot for this order, when available
          oneOf:
            - $ref: "#/components/schemas/Address"
            - type: "null"
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        fulfillments:
          type: array
          items:
            $ref: "#/components/schemas/Fulfillment"
    OrderList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: integer }
    VendorOrderItem:
      allOf:
        - $ref: "#/components/schemas/OrderItem"
        - type: object
          required: [orderStatus, orderCreatedAt, fulfillmentStatus]
          properties:
            orderStatus:
              $ref: "#/components/schemas/OrderStatus"
            orderCreatedAt: { type: string, format: date-time }
            fulfillmentStatus:
              $ref: "#/components/schemas/FulfillmentStatus"
    CheckoutInput:
      type: object
      required: [addressId, paymentMethod]
      properties:
        promoCode: { type: string, minLength: 1 }
        addressId: { type: integer }
        paymentMethod:
          $ref: "#/components/schemas/OrderPaymentMethod"
        momoPhone:
          description: Required when paymentMethod is "momo" — the wallet phone number to charge, e.g. "0244000000"
          type: string
          minLength: 9
        momoProvider:
          description: Required when paymentMethod is "momo"
          $ref: "#/components/schemas/MomoProvider"
    Address:
      type: object
      required: [id, buyerUserId, fullName, phone, line1, city, isDefault, createdAt]
      properties:
        id: { type: integer }
        buyerUserId: { type: integer }
        label: { type: ["string", "null"] }
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        city: { type: string }
        region: { type: ["string", "null"] }
        digitalAddress: { type: ["string", "null"] }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    FulfillmentStatus:
      type: string
      enum: [unfulfilled, packed, shipped, delivered]
    Fulfillment:
      type: object
      required: [id, orderId, vendorId, status]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        vendorId: { type: integer }
        status:
          $ref: "#/components/schemas/FulfillmentStatus"
        packedAt: { type: ["string", "null"], format: date-time }
        shippedAt: { type: ["string", "null"], format: date-time }
        deliveredAt: { type: ["string", "null"], format: date-time }
...
    VendorOrderItemList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VendorOrderItem"
        total:
          type: integer
...